requirements.txt
pytest-playwright==0.3.2